Refactored Image Bar
