--- src/main/java/entity/MidMenuBar/ColorButtonsBar/ColorButtonsFactory.java
@@ -1,4 +1,0 @@
-package entity.MidMenuBar.ColorButtonsBar;
-
-public interface ColorButtonsFactory {
-}
