--- src/main/java/entity/MidMenuBar/ColorButtonsBar/ColorWheel.java
@@ -1,4 +1,0 @@
-package entity.MidMenuBar.ColorButtonsBar;
-
-public class ColorWheel {
-}
