--- src/main/java/entity/MidMenuBar/SelectionToolButton.java
@@ -1,4 +1,0 @@
-package entity.MidMenuBar;
-
-public class SelectionToolButton {
-}
