+++ src/main/java/view/MidMenuBar/ChangeBrushButton.java
@@ -1,4 +1,4 @@
-package entity.MidMenuBar;
+package view.MidMenuBar;
 
 public class ChangeBrushButton {
 }
