+++ src/main/java/view/MidMenuBar/ColorButtonsBar/ColorButtonGenerator.java
@@ -1,0 +1,4 @@
+package view.MidMenuBar.ColorButtonsBar;
+
+public class ColorButtonGenerator {
+}
