+++ src/main/java/view/MidMenuBar/ColorButtonsBar/ColorButtonsFactory.java
@@ -1,4 +1,4 @@
-package view.MidMenuBar;
+package view.MidMenuBar.ColorButtonsBar;
 
 public interface ColorButtonsFactory {
 }
