--- src/main/java/view/MidMenuBar/CropButton.java
@@ -1,4 +1,0 @@
-package view.MidMenuBar;
-
-public class CropButton {
-}
