+++ src/main/java/view/MidMenuBar/ImageBar/CropButton.java
@@ -1,0 +1,4 @@
+package view.MidMenuBar.ImageBar;
+
+public class CropButton {
+}
