+++ src/main/java/view/MidMenuBar/ImageBar/ResizeButton.java
@@ -1,0 +1,4 @@
+package view.MidMenuBar.ImageBar;
+
+public class ResizeButton {
+}
