+++ src/main/java/view/MidMenuBar/ImageBar/RotateButton.java
@@ -1,0 +1,4 @@
+package view.MidMenuBar.ImageBar;
+
+public class RotateButton {
+}
