--- src/main/java/view/MidMenuBar/ResizeButton.java
@@ -1,4 +1,0 @@
-package view.MidMenuBar;
-
-public class ResizeButton {
-}
