--- src/main/java/view/MidMenuBar/RotateButton.java
@@ -1,4 +1,0 @@
-package view.MidMenuBar;
-
-public class RotateButton {
-}
