+++ src/main/java/view/TopMenuBar/ViewMenu/ZoomOutButton.java
@@ -1,6 +1,5 @@
 package view.TopMenuBar.ViewMenu;
 
-import entity.CanvasState;
 import entity.DrawingCanvas;
 
 import javax.swing.*;
